build.gradle: fix custom Eclipse output directories
Workaround introduced by 764e9aa4b077 to keep a single Eclipse output
directory, set by 'defaultOutputDir', doesn't work anymore when using
Gradle 6.4. Eclipse 20.03 does require separated output folders because
the test sourceSet is marked with the attribute 'test' in the .classpath
file now, and it will fail with:

    Test source folder 'test' in project 'clara-java' must have an
    output folder that is not also used for main sources

But the base output directory is not based in the default output
directory. If 'source.output' is not forced to be 'defaultOutputDir', to
avoid the Eclipse error, then Gradle will set the output directores to:

    - default:  build/eclipse/
    - main:     bin/main/
    - test:     bin/test/

So it needs to be fixed manually. All of them output folders cannot be
the same nor nested inside each other, or Eclipse will not build the
project.

This commits ensures the output directores are set to:

    - default:  build/eclipse/default/
    - main:     build/eclipse/main/
    - test:     build/eclipse/test/

This is, just like the defaults, but with 'build/eclipse' instead of
'bin'.

Also use a relative path for 'defaultOutputDir' instead of the full
project path (just like 'source.output', which must be a relative
directory).

See related Gradle issues 3825 and 3839.