query: refactor DpeQuery class
Make clear that it is for requesting and filtering the DPE reports (JSON).

The query will parse the JSON report to obtain a stream of the
component-specific JSON objects, the stream will be filtered by the query
filter, and then the resulting JSON object stream will be parsed into a
stream of the required data type for derived classes to collect into the
query return value.