build.gradle: fix running Spock integration tests
Spock doesn't support the Junit `@Tag` annotation to filter tests by
category. For the integration tests, this means that if the Junit Platform
is configured to select the tag 'integration' in the integrationTest task,
none of the Spock integration tests will be selected and they will not run.

Fix by using the Spock config file to include or exclude integration tests
by using a custom annotation, and remove the configuration of the Junit
Platform to include just the 'integration' tag in the integrationTest task.

This means:

- Unit tests: none of them will be in Java, and Spock will select those
  that are not annotated with `@Integration`. Keep the exclude
  'integration' tag setting for the Junit Platform, to ensure that the
  integration tests that are still in Java will not run.

- Integration tests: they will be a mix of Spock and Java (the tests still
  in Java mostly for testing performance). Spock will only select those
  annotated with `@Integration`. By not configuring the Junit Platform to
  include the 'integration' tag, all Java tests will also run. And since
  the only Java tests that will be left are those annotated with the
  'integration' tag, then they will run as expected.

Use Gradle test fixtures to declare the custom `@Integration` annotation in
just one subproject, and let the other subprojects use it. The clara-msg
subproject seems the best place to put it.