build.gradle: workaround for deprecation warning in Protobuf plugin
The Protobuf plugin extracts proto files from dependencies into the
build directory. But with the new testFixtures source set introduced in
6da7794f (build.gradle: fix running Spock integration tests, 2021-02-26)
now both the test and testFixtures sets depend on main, and test depends
on testFixtures, and something is going on and the proto files are being
extracted twice:

    Copying or archiving duplicate paths with the default duplicates strategy has been deprecated.
    This is scheduled to be removed in Gradle 7.0.
    Duplicate path: "org/jlab/clara/msg/data/MetaData.proto".
    Explicitly set the duplicates strategy to 'DuplicatesStrategy.INCLUDE' if you want to allow duplicate paths.

Since the extraction of proto files in the dependencies is meant to add
them to the protoc path when compiling proto files in the current source
set, the extraction can be just disabled for the tests (the tests do no
compile their own proto files), until fixed upstream.