Try to mitigate test failures in the CI
It seems that the CI is too slow or does not have enough cores, and ZMQ
starts dropping messages (looks like due to the subscribe propagation
taking too long).

Try to add more time in order to wait for control messages from the
proxy to check if a connection/subscription is "almost probably ready".
Also add extra delay after creating the connection/subscription to
give more time for the subscribe propagation.

And add more time to wait for messages in the publishers test, which is
needed because the tests take longer to finish in the CI now.