tests: clarify that proxy/registrar wrappers can fail
The proxy/registrar wrapper can fail to be created if there is already a
proxy/registrar running on the local node, external to the tests, and
the exception will just be reported and the tests will continue.

This is to support such external proxy/registrar when running the tests,
if needed.