Refactor registrar service
Be explicit that each request creates a response, instead of having only some of
the requests filling the result data, and then the default case returning
directly instead of setting a response data. Prefer to be clear instead of
saving a few lines: each case explicitly creates its response object.