style: don't be so strict over the line length
Do try to keep the lines under 100 characters, but don't complain if
some line is just a few characters longer. Most of the time it looks
better at 101 or 102 characters long than breaking the line to keep it
under 100. So, move the line length limit up to 108 to allow those extra
characters when needed. 8% extra seems a good compromise (for example,
Black allows up to 88 characters, 10%, for Python).