registration: remove filter by topic parts
The artificial separation of topic into parts doesn't matter at all for
ZeroMQ, so it's not really important as a registration filter.

Do replace it with a "prefix" filter, this is, registered actors
(publishers and subscribers) whose topic starts with the given prefix,
which is more useful. While it is the same as normal topic matching for
finding publishers, for subscribers it forces prefix matching (since the
find request uses reverse matching for subscribers), which allows to
filters subscribers with the same top-hierarchy of topics.