registration: refactor registrar requests
No point in having two sets of duplicated request strings, for publishers
and subscribers, when the registration data already contains the type of
the request.

Use the registration data to get the type of the request and select the
proper publishers or subscribers database, instead of relying on the
duplicated per-type request strings.

Also make clear that the request string is a registration action and not a
topic, which made it confusing with the actors pub-sub topics.