std: fix standard orchestrator benchmark report
Broken by 0d027f50 (Migrate some classes to records, 2022-01-04), which
refactored ServiceInfo as a record. That resulted in its equals() and
hashCode() methods now using the classpath field when originally it was
ignored.

But the Benchmark class used the ServiceInfo objects obtained from
ApplicationInfo as keys of its internal map, and to access the values it
was constructing dummy ServiceInfo objects from a ServiceName, with an
empty classpath. Since the keys are records now, even if the services
are present in the map, the key is not found due to originally being
stored with a full classpath and the getter using a key with an empty
classpath.

Fix by using a proper custom class for the keys, with only the required
fields, ignoring the classpath.